added LibVocabCard component
