+++ src/components/LibVocabulary.js
@@ -2,8 +2,6 @@ import {useState, useEffect} from "react"
 
 export default function LibVocabulary() {
 
-    const [showBack, setShowBack] = useState(false)
-    
     const [vocab, setVocab] = useState([])
     
     useEffect( ()=>{
@@ -13,25 +11,10 @@ export default function LibVocabulary() {
         }, []
     )
 
-    const handleFlip = () => {
-        setShowBack(!showBack)
-    }
 
     
     const wordComponent = vocab.map( word => {
-        return(
-            <div>
-                {showBack ?
-                    <div onClick={handleFlip} class="hover:cursor-pointer bg-[#E8E4DE] rounded-full shadow-lg backdrop-blur-lg dark:bg-slate-800/50 w-52 h-52 m-10 p-10">
-                        <p class="text-[#292929] text-xl text-center font-semibold">{word.definition}</p>
-                    </div> 
-                    :
-                    <div onClick={handleFlip} class="hover:cursor-pointer bg-[#292929] rounded-full shadow-lg backdrop-blur-lg dark:bg-slate-800/50 w-52 h-52 m-10 p-10">
-                        <h1 class="text-[#E8E4DE] text-xl text-center font-semibold">{word.term}</h1>
-                    </div>
-                }
-            </div>
-        )
+        
     })
